+++ Panels/Panel.js
@@ -126,9 +126,11 @@ const PanelBase = kind({
 	handlers: {
 		onScroll: handle(
 			forward('onScroll'),
-			({currentTarget}) => {
-				currentTarget.scrollTop = 0;
-				currentTarget.scrollLeft = 0;
+			({currentTarget, eventTarget}) => {
+				if (currentTarget === eventTarget) {
+					currentTarget.scrollTop = 0;
+					currentTarget.scrollLeft = 0;
+				}
 			}
 		)
 	},
@@ -185,6 +187,7 @@ const PanelBase = kind({
 				<div className={css.header} id={headerId}>
 					<ComponentOverride
 						component={header}
+						data-index={rest['data-index']}
 						entering={entering}
 						subtitleId={subtitleId}
 						titleId={titleId}