+++ samples/qa-a11y/src/views/RangePicker.js
@@ -59,7 +59,7 @@ const RangePickerView = () => (
 			/>
 		</Section>
 
-		<SpotlightContainerSection className={appCss.marginTop} title="Vertical" vertical>
+		<SpotlightContainerSection className={appCss.marginTop} horizontal title="Vertical">
 			<RangePicker
 				alt="Vertical"
 				defaultValue={0}
@@ -162,7 +162,7 @@ const RangePickerView = () => (
 			/>
 		</Section>
 
-		<SpotlightContainerSection className={appCss.marginTop} title="Vertical" vertical>
+		<SpotlightContainerSection className={appCss.marginTop} horizontal title="Vertical">
 			<RangePicker
 				alt="Vertical"
 				decrementAriaLabel="This is a Label 9."
@@ -270,7 +270,7 @@ const RangePickerView = () => (
 			/>
 		</Section>
 
-		<SpotlightContainerSection className={appCss.marginTop} title="Vertical" vertical>
+		<SpotlightContainerSection className={appCss.marginTop} horizontal title="Vertical">
 			<RangePicker
 				alt="Vertical"
 				aria-label="This is a Label 21."