+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [1.5.0] - 2021-02-09
+
+### Added
+
+- `sandstone/Item` prop `data-webos-voice-labels` when `label` is used
+
+### Fixed
+
+- `sandstone/Alert` to read out properly after closing it in a `sandstone/PopupTabLayout`
+- `sandstone/FlexiblePopupPanels` padding in RTL locales
+- `sandstone/Heading` `font-style` to use oblique font instead of fake `italic`
+- `sandstone/Input` to not have initial focus with pointer when `type` prop is `'number'` or `'passwordnumber'`
+- `sandstone/Panel` to not reset scroll position by events from others
+- `sandstone/Panels.Header` to not show back button in the first panel
+- `sandstone/VideoPlayer.Video` to not start a new play before another one completes
+
 ## [1.4.6] - 2021-01-29
 
 ### Fixed